requirements.txt
Django==4.2.20
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
Pillow==11.1.0 